Extract content and components into separate files
